fix: DateTimeInput displaying empty label
The DateTimeInput displays a label with only the format while it should
not display any label when the label property is not defined.

Converted MarkedAsRequiredLabel into hook to be able to return label as
is when no required mark should be added.
